+++ gantry/__main__.py
@@ -592,5 +592,68 @@ def set_gh_token(
     )
 
 
+@main.group(**_CLICK_GROUP_DEFAULTS)
+def cluster():
+    """
+    Get information on Beaker clusters.
+    """
+
+
+@cluster.command(name="list", **_CLICK_COMMAND_DEFAULTS)  # type: ignore
+@click.option(
+    "--cloud",
+    is_flag=True,
+    help="""Only show cloud clusters.""",
+)
+def list_clusters(cloud: bool = False):
+    """
+    List available clusters.
+
+    By default only on-premise clusters are displayed.
+    """
+    beaker = Beaker.from_env(session=True)
+    clusters = [c for c in beaker.cluster.list() if c.is_cloud == cloud]
+    for cluster in clusters:
+        icon = "☁️" if cluster.is_cloud else "🏠"
+        print(f"{icon} [b magenta]{cluster.full_name}[/]")
+        for node in sorted(beaker.cluster.nodes(cluster), key=lambda node: node.hostname):
+            print(
+                f"   [i cyan]{node.hostname}[/] - "
+                f"CPUs: {node.limits.cpu_count}, "
+                f"GPUs: {node.limits.gpu_count or 0} {'x' if node.limits.gpu_type else ''} {node.limits.gpu_type or ''}"
+            )
+        if cluster.node_spec is not None:
+            limits = cluster.node_spec
+            print(
+                f"  CPUs: {limits.cpu_count}, "
+                f"GPUs: {limits.gpu_count or 0} {'x' if limits.gpu_type else ''} {limits.gpu_type or ''}"
+            )
+
+
+@cluster.command(name="util", **_CLICK_COMMAND_DEFAULTS)  # type: ignore
+@click.argument("cluster", nargs=1, required=True, type=str)
+def cluster_util(cluster: str):
+    """
+    Get the current status and utilization for a cluster.
+    """
+    beaker = Beaker.from_env(session=True)
+    cluster_util = beaker.cluster.utilization(cluster)
+    cluster = cluster_util.cluster
+    icon = "☁️" if cluster.is_cloud else "🏠"
+    print(
+        f"{icon} [b magenta]{cluster.full_name}[/]\n\n"
+        f"running jobs: {cluster_util.running_jobs} ({cluster_util.running_preemptible_jobs} preemptible)\n"
+        f"queued jobs: {cluster_util.queued_jobs}"
+    )
+    if cluster_util.nodes:
+        print("nodes:")
+    for node in sorted(cluster_util.nodes, key=lambda n: n.hostname):
+        print(
+            f"  [i cyan]{node.hostname}[/] - {node.running_jobs} jobs ({node.running_preemptible_jobs} preemptible)\n"
+            f"    CPUs free: [{'green' if node.free.cpu_count else 'red'}]{node.free.cpu_count} / {node.limits.cpu_count}[/]\n"
+            f"    GPUs free: [{'green' if node.free.gpu_count else 'red'}]{node.free.gpu_count or 0} / {node.limits.gpu_count}[/] {node.free.gpu_type or ''}\n"
+        )
+
+
 if __name__ == "__main__":
     main()